Prepare for 0.3.0 release
changes listed in changelog

+++ src/key_handler/mod.rs
@@ -24,13 +24,13 @@ const X: KeyModifiers = KeyModifiers::NONE;
 const S: KeyModifiers = KeyModifiers::SHIFT;
 const C: KeyModifiers = KeyModifiers::CONTROL;
 
-const SEEK_SMALL: u64 = 5;
-const SEEK_LARGE: u64 = 30;
+const SEEK_SMALL: f64 = 5.0;
+const SEEK_LARGE: f64 = 30.0;
 const SCROLL_MID: usize = 5;
 const SCROLL_XTRA: usize = 20;
 const SIDEBAR_INCREMENT: isize = 1;
 
-#[derive(PartialEq, Eq)]
+#[derive(PartialEq)]
 pub enum Action {
     // Player Controls
     Play(usize),
@@ -38,8 +38,8 @@ pub enum Action {
     TogglePlayback,
     PlayNext,
     PlayPrev,
-    SeekForward(u64),
-    SeekBack(u64),
+    SeekForward(f64),
+    SeekBack(f64),
 
     // Queue & Playlist Actions
     QueueSong,
